# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mky fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FitnessGoal = (typeof FitnessGoal)[keyof typeof FitnessGoal];

export const FitnessGoal = {
  lose_weight: "lose_weight",
  maintain: "maintain",
  build_muscle: "build_muscle",
  improve_endurance: "improve_endurance",
} as const;

export type ActivityLevel = (typeof ActivityLevel)[keyof typeof ActivityLevel];

export const ActivityLevel = {
  sedentary: "sedentary",
  light: "light",
  moderate: "moderate",
  active: "active",
  very_active: "very_active",
} as const;

export type Sex = (typeof Sex)[keyof typeof Sex];

export const Sex = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface Profile {
  id: string;
  displayName: string;
  sex: Sex;
  age: number;
  heightCm: number;
  weightKg: number;
  goal: FitnessGoal;
  activityLevel: ActivityLevel;
  targetWeightKg?: number | null;
}

export interface UpdateProfileRequest {
  displayName: string;
  sex: Sex;
  age: number;
  heightCm: number;
  weightKg: number;
  goal: FitnessGoal;
  activityLevel: ActivityLevel;
  targetWeightKg?: number | null;
}

export interface CalorieRecommendation {
  bmr: number;
  tdee: number;
  targetCalories: number;
  proteinGrams: number;
  carbGrams: number;
  fatGrams: number;
  explanation: string;
}

export interface Exercise {
  id: string;
  name: string;
  sets: number;
  reps: number;
  weightKg?: number | null;
  durationSeconds?: number | null;
  notes?: string | null;
}

export interface ExerciseInput {
  name: string;
  sets: number;
  reps: number;
  weightKg?: number | null;
  durationSeconds?: number | null;
  notes?: string | null;
}

export interface Workout {
  id: string;
  name: string;
  description?: string | null;
  /** e.g. push, pull, legs, cardio, full body */
  category: string;
  estimatedMinutes: number;
  exercises: Exercise[];
  createdAt: string;
}

export interface CreateWorkoutRequest {
  name: string;
  description?: string | null;
  category: string;
  estimatedMinutes: number;
  exercises: ExerciseInput[];
}

export interface UpdateWorkoutRequest {
  name: string;
  description?: string | null;
  category: string;
  estimatedMinutes: number;
  exercises: ExerciseInput[];
}

export interface Session {
  id: string;
  workoutId: string;
  workoutName: string;
  completedAt: string;
  durationMinutes: number;
  caloriesBurned?: number | null;
  notes?: string | null;
}

export interface CompleteSessionRequest {
  workoutId: string;
  durationMinutes: number;
  caloriesBurned?: number | null;
  notes?: string | null;
}

export type Weekday = (typeof Weekday)[keyof typeof Weekday];

export const Weekday = {
  mon: "mon",
  tue: "tue",
  wed: "wed",
  thu: "thu",
  fri: "fri",
  sat: "sat",
  sun: "sun",
} as const;

export interface ScheduleEntry {
  id: string;
  weekday: Weekday;
  /** HH:MM 24-hour format */
  timeOfDay: string;
  workoutId: string;
  workoutName: string;
}

export interface ScheduleEntryInput {
  weekday: Weekday;
  timeOfDay: string;
  workoutId: string;
}

export interface UpsertScheduleRequest {
  entries: ScheduleEntryInput[];
}

export type ReminderKind = (typeof ReminderKind)[keyof typeof ReminderKind];

export const ReminderKind = {
  missed: "missed",
  upcoming: "upcoming",
} as const;

export interface Reminder {
  id: string;
  kind: ReminderKind;
  message: string;
  workoutId: string;
  workoutName: string;
  scheduledFor: string;
}

export type StreakWeekHistoryItem = {
  date: string;
  completed: boolean;
};

export interface Streak {
  currentDays: number;
  bestDays: number;
  lastActiveDate?: string | null;
  /** Last 7 days, oldest to newest */
  weekHistory: StreakWeekHistoryItem[];
}

export type DashboardWeeklyStats = {
  sessionsThisWeek: number;
  minutesThisWeek: number;
  caloriesThisWeek: number;
  goalSessions: number;
};

export interface Dashboard {
  greeting: string;
  todayDate: string;
  todaysWorkout?: ScheduleEntry | null;
  todayCompleted: boolean;
  streak: Streak;
  weeklyStats: DashboardWeeklyStats;
  recentSessions: Session[];
  activeReminders: Reminder[];
}

export type TrainerChatMessageRole =
  (typeof TrainerChatMessageRole)[keyof typeof TrainerChatMessageRole];

export const TrainerChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface TrainerChatMessage {
  role: TrainerChatMessageRole;
  content: string;
}

export interface TrainerChatRequest {
  message: string;
  history?: TrainerChatMessage[];
}

export interface TrainerReply {
  reply: string;
}

export type ListSessionsParams = {
  limit?: number;
};