# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mky fitness tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List all workout routines
      responses:
        "200":
          description: Workouts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create a workout routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{workoutId}:
    parameters:
      - name: workoutId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get a workout routine
      responses:
        "200":
          description: Workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    put:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update a workout routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkoutRequest"
      responses:
        "200":
          description: Updated workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete a workout routine
      responses:
        "204":
          description: Deleted
...
  /trainer/chat:
    post:
      operationId: trainerChat
      tags: [trainer]
      summary: Chat with the AI personal trainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerReply"
components:
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sets:
          type: integer
        reps:
          type: integer
        weightKg:
          type: number
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, name, sets, reps]
    ExerciseInput:
      type: object
      properties:
        name:
          type: string
        sets:
          type: integer
        reps:
          type: integer
        weightKg:
          type: number
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, sets, reps]
    Workout:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          description: e.g. push, pull, legs, cardio, full body
        estimatedMinutes:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, estimatedMinutes, exercises, createdAt]
    CreateWorkoutRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        estimatedMinutes:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseInput"
      required: [name, category, estimatedMinutes, exercises]
    UpdateWorkoutRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        estimatedMinutes:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseInput"
      required: [name, category, estimatedMinutes, exercises]
...
    TrainerChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required: [role, content]
    TrainerChatRequest:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/TrainerChatMessage"
      required: [message]
    TrainerReply:
      type: object
      properties:
        reply:
          type: string
      required: [reply]