# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mky fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface Workout {
  id: string;
  name: string;
  description?: string | null;
  /** e.g. push, pull, legs, cardio, full body */
  category: string;
  estimatedMinutes: number;
  exercises: Exercise[];
  createdAt: Date;
}